+++ empire/server/core/db/models.py
@@ -28,7 +28,10 @@ from sqlalchemy.ext.hybrid import hybrid_property
 from sqlalchemy.orm import Mapped, declarative_base, deferred, relationship
 from sqlalchemy_utc import UtcDateTime, utcnow
 
-from empire.server.core.config import PluginAutoExecuteConfig, empire_config
+from empire.server.core.config.config_manager import (
+    PluginAutoExecuteConfig,
+    empire_config,
+)
 from empire.server.core.module_models import EmpireAuthor
 from empire.server.utils.datetime_util import is_stale