+++ empire/server/utils/log_util.py
@@ -3,7 +3,7 @@ import os
 import pwd
 from pathlib import Path
 
-from empire.server.core.config import empire_config
+from empire.server.core.config.config_manager import empire_config
 
 LOG_FORMAT = "%(asctime)s [%(filename)s:%(lineno)d] [%(levelname)s]: %(message)s "
 SIMPLE_LOG_FORMAT = "[%(levelname)s]: %(message)s "