+++ empire/test/test_logs.py
@@ -42,7 +42,7 @@ def test_extended_log_format(monkeypatch):
     monkeypatch.setattr("empire.server.server.empire", MagicMock())
 
     from empire import arguments
-    from empire.server.core.config import EmpireConfig
+    from empire.server.core.config.config_manager import EmpireConfig
     from empire.server.server import setup_logging
     from empire.server.utils.log_util import LOG_FORMAT, ColorFormatter
 
@@ -73,7 +73,7 @@ def test_log_level_by_config(monkeypatch):
     monkeypatch.setattr("empire.server.server.empire", MagicMock())
 
     from empire import arguments
-    from empire.server.core.config import EmpireConfig
+    from empire.server.core.config.config_manager import EmpireConfig
     from empire.server.server import setup_logging
 
     test_config = load_test_config()
@@ -155,7 +155,7 @@ def test_log_file_not_owned_by_root(monkeypatch):
     monkeypatch.setattr("empire.server.server.empire", MagicMock())
 
     from empire import arguments
-    from empire.server.core.config import EmpireConfig
+    from empire.server.core.config.config_manager import EmpireConfig
     from empire.server.server import setup_logging
 
     test_config = load_test_config()